assignment_badge_converter_backend/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
openai==1.51.2
httpx==0.27.2
